Simplified Spec tests via ABC
